refactor surface-specific modules
