+++ package-files.test.ts
@@ -42,6 +42,10 @@ describe("package manifest files", () => {
         "catalog/build-provider.ts",
         "provider-catalog.ts",
         "provider-discovery.ts",
+        "web-search/credentials.ts",
+        "web-search/results.ts",
+        "image/request.ts",
+        "image/response.ts",
         "image-generation-provider.ts",
         "web-search.ts",
         "onboard.ts",
