+++ package.json
@@ -40,6 +40,10 @@
     "provider-catalog.ts",
     "provider-discovery.ts",
     "nanogpt-errors.ts",
+    "web-search/credentials.ts",
+    "web-search/results.ts",
+    "image/request.ts",
+    "image/response.ts",
     "image-generation-provider.ts",
     "web-search.ts",
     "onboard.ts",
